requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
numpy==1.22.3
pynput==1.7.6
pyobjc-core==8.4.1
pyobjc-framework-ApplicationServices==8.4.1
pyobjc-framework-Cocoa==8.4.1
pyobjc-framework-Quartz==8.4.1
PyQt6-Qt6==6.2.3
PyQt6-WebEngine==6.3.0
PyQt6-WebEngine-Qt6==6.3.0
requests==2.27.1
six==1.16.0
urllib3==1.26.9
youtube-dl==2021.12.17
